[FIX] Reorganise and reduce size of function
