+++ src/struct_filler/identify_cmd.c
@@ -1,58 +1,40 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   struct_fill.c                                      :+:      :+:    :+:   */
+/*   identify_cmd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jhurpy <jhurpy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/12 19:46:32 by whendrik          #+#    #+#             */
-/*   Updated: 2024/01/15 00:09:00 by jhurpy           ###   ########.fr       */
+/*   Created: 2024/01/15 16:25:44 by jhurpy            #+#    #+#             */
+/*   Updated: 2024/01/15 16:51:39 by jhurpy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-void	init_cmd(t_cmd *cmd, int j, t_tokens *tokens)
+static void	init_cmd(t_cmd *cmd, int j, t_tokens *tokens)
 {
 	if (j < (tokens->pipe_count) && tokens->pipe_count > 0)
 		cmd->pipe_out = TRUE;
-	else
-		cmd->pipe_out = FALSE;
 	if (j > 0)
 		cmd->pipe_in = TRUE;
-	else
-		cmd->pipe_in = FALSE;
 	if (tokens->heredoc_count[j] != 0 && tokens->heredoc_count[j] > 0
 		&& tokens->infile_count[j] == 0)
 		cmd->here_doc_in = TRUE;
-	else
-		cmd->here_doc_in = FALSE;
 	if ((tokens->infile_count[j] != 0 && tokens->infile_count[j] > 0)
 		|| cmd->here_doc_in == TRUE)
 		cmd->file_in = TRUE;
-	else
-		cmd->file_in = FALSE;
-	if (tokens->append_count[j] != 0 && tokens->append_count[j] > 0)						/*Check if its only if it appears last*/
+	if (tokens->append_count[j] != 0 && tokens->append_count[j] > 0)
 		cmd->append = TRUE;
-	else
-		cmd->append = FALSE;
 	if ((tokens->outfile_count[j] != 0 && tokens->outfile_count[j] > 0)
 		|| cmd->append == TRUE)
 		cmd->file_out = TRUE;
-	else
-		cmd->file_out = FALSE;
-	cmd->here_doc_fd = -1;
-	cmd->cmd = NULL;
-	cmd->limiters = NULL;
-	cmd->infiles = NULL;
-	cmd->outfiles = NULL;
 }
 
-void	mallocer(t_cmd *cmd, t_tokens *tokens, int j)
+static void	malloc_cmd(t_cmd *cmd, t_tokens *tokens, int j)
 {
-	if (cmd->cmd == NULL)
-		cmd->cmd = (char **)ft_calloc(sizeof(char *),
-				((tokens->arg_count[j]) + 1));
+	cmd->cmd = (char **)ft_calloc(sizeof(char *),
+			((tokens->arg_count[j]) + 1));
 	cmd->limiters = (char **)ft_calloc(sizeof(char *),
 			(tokens->heredoc_count[j] + 1));
 	cmd->nb_heredocs = tokens->heredoc_count[j];
@@ -62,7 +44,7 @@ void	mallocer(t_cmd *cmd, t_tokens *tokens, int j)
 			(tokens->outfile_count[j] + tokens->append_count[j] + 1));
 }
 
-void	sort_rdrt(t_cmd *cmd, char *token, char *next_token)
+static void	sort_rdrt(t_cmd *cmd, char *token, char *next_token)
 {
 	int	i;
 
@@ -87,7 +69,7 @@ void	sort_rdrt(t_cmd *cmd, char *token, char *next_token)
 	}
 }
 
-void	find_last_rdrt(t_cmd *cmd, t_tokens *tokens, int i)
+static void	find_last_rdrt(t_cmd *cmd, t_tokens *tokens, int i)
 {
 	while (tokens->token_type[i] != e_pipe && i != 0)
 	{
@@ -105,20 +87,19 @@ void	find_last_rdrt(t_cmd *cmd, t_tokens *tokens, int i)
 	}
 }
 
-void	identify_2(t_cmd *cmd, t_tokens *tokens, int j, int *i)
+void	identify_cmd(t_cmd *cmd, t_tokens *tokens, int j, int *i)
 {
 	int	x;
 
 	x = 0;
 	init_cmd(cmd, j, tokens);
-	mallocer(cmd, tokens, j);
+	malloc_cmd(cmd, tokens, j);
 	while (*i < tokens->token_count && tokens->token_type[*i] != e_pipe)
 	{
 		if (tokens->token_type[*i] == e_rdrt)
 		{
 			sort_rdrt(cmd, tokens->tokens[*i], tokens->tokens[*i + 1]);
 			*i += 2;
-																/*Only works bc token_syntax makes sure no pipe or rdrt comes after*/
 		}
 		else if (tokens->token_type[*i] == e_argument)
 		{
@@ -128,23 +109,4 @@ void	identify_2(t_cmd *cmd, t_tokens *tokens, int j, int *i)
 		}
 	}
 	find_last_rdrt(cmd, tokens, *i - 1);
-}
-
-bool	struct_fill(t_tokens *tokens, t_data *data)						/*TOO MANY FUNCTION IN FILE*/
-{
-	int		i;
-	int		j;
-	t_cmd	*cmd_struct;
-
-	cmd_struct = (t_cmd *)malloc(sizeof(t_cmd) * (tokens->pipe_count + 2));
-	i = 0;
-	j = 0;
-	while (j < tokens->pipe_count + 1)
-	{
-		identify_2(&cmd_struct[j], tokens, j, &i);
-		i += 1; 														/*So its not stuck on pipe*/
-		j++;
-	}
-	data->cmd = cmd_struct;
-	return (1);
 }
